Push SortThree

+++ src/main/java/viikko1/SortThree.java
@@ -3,9 +3,9 @@ package viikko1;
 
 public class SortThree {
     public static String ascending(int a, int b, int c) {
-        double max = Math.max(a, Math.max(b, c));
-        double min = Math.min(a, Math.min(b, c));
-        double mid = a+b+c-max-min;
+        int max = Math.max(a, Math.max(b, c));
+        int min = Math.min(a, Math.min(b, c));
+        int mid = a+b+c-max-min;
         
         return ""+min+","+mid+","+max;
     }
